src/main/java/lpsvisualizer/websocket/PositionWebSocketHandler.java
package lpsvisualizer.websocket;


import com.fasterxml.jackson.databind.ObjectMapper;
import lpsvisualizer.entity.DisplayablePosition;
import org.springframework.web.socket.CloseStatus;
import org.springframework.web.socket.TextMessage;
import org.springframework.web.socket.WebSocketSession;
import org.springframework.web.socket.handler.TextWebSocketHandler;

import java.io.IOException;
import java.util.List;
import java.util.concurrent.CopyOnWriteArrayList;

/**
 * @since 26.11.2024
 */
public class PositionWebSocketHandler extends TextWebSocketHandler {

    private final CopyOnWriteArrayList<WebSocketSession> sessions = new CopyOnWriteArrayList<>();

    private final ObjectMapper mapper = new ObjectMapper();

    @Override
    public void afterConnectionEstablished(WebSocketSession session) throws Exception {
        sessions.add(session);
    }

    @Override
    public void afterConnectionClosed(WebSocketSession session, CloseStatus status) throws Exception {
        sessions.remove(session);
    }

    public void sendPositionsToClients(List<DisplayablePosition> positions) {
        for (WebSocketSession session : sessions) {
            if (session.isOpen()) {
                try {
                    session.sendMessage(new TextMessage(mapper.writeValueAsString(positions)));
                } catch (IOException e) {
                    e.printStackTrace();
                }
            }
        }
    }
}
